+++ src/Google/Protobuf/EnumDescriptorProto.php
@@ -1,17 +1,19 @@
 <?php
 /**
- * Auto generated from descriptor.proto at 2016-06-26 19:19:31
+ * Auto generated from descriptor.proto at 2016-07-08 14:22:37
+ *
+ * google.protobuf package
  */
 
-namespace Google\Protobuf\Compiler {
+namespace Google\Protobuf {
 /**
- * ServiceDescriptorProto message
+ * EnumDescriptorProto message
  */
-class ServiceDescriptorProto extends \ProtobufMessage
+class EnumDescriptorProto extends \ProtobufMessage
 {
     /* Field index constants */
     const NAME = 1;
-    const METHOD = 2;
+    const VALUE = 2;
     const OPTIONS = 3;
 
     /* @var array Field descriptors */
@@ -19,17 +21,17 @@ class ServiceDescriptorProto extends \ProtobufMessage
         self::NAME => array(
             'name' => 'name',
             'required' => false,
-            'type' => 7,
+            'type' => \ProtobufMessage::PB_TYPE_STRING,
         ),
-        self::METHOD => array(
-            'name' => 'method',
+        self::VALUE => array(
+            'name' => 'value',
             'repeated' => true,
-            'type' => '\Google\Protobuf\Compiler\MethodDescriptorProto'
+            'type' => '\Google\Protobuf\EnumValueDescriptorProto'
         ),
         self::OPTIONS => array(
             'name' => 'options',
             'required' => false,
-            'type' => '\Google\Protobuf\Compiler\ServiceOptions'
+            'type' => '\Google\Protobuf\EnumOptions'
         ),
     );
 
@@ -49,7 +51,7 @@ class ServiceDescriptorProto extends \ProtobufMessage
     public function reset()
     {
         $this->values[self::NAME] = null;
-        $this->values[self::METHOD] = array();
+        $this->values[self::VALUE] = array();
         $this->values[self::OPTIONS] = null;
     }
 
@@ -86,77 +88,77 @@ class ServiceDescriptorProto extends \ProtobufMessage
     }
 
     /**
-     * Appends value to 'method' list
+     * Appends value to 'value' list
      *
-     * @param \Google\Protobuf\Compiler\MethodDescriptorProto $value Value to append
-     *
-     * @return null
-     */
-    public function appendMethod(\Google\Protobuf\Compiler\MethodDescriptorProto $value)
-    {
-        return $this->append(self::METHOD, $value);
-    }
-
-    /**
-     * Clears 'method' list
+     * @param \Google\Protobuf\EnumValueDescriptorProto $value Value to append
      *
      * @return null
      */
-    public function clearMethod()
+    public function appendValue(\Google\Protobuf\EnumValueDescriptorProto $value)
     {
-        return $this->clear(self::METHOD);
+        return $this->append(self::VALUE, $value);
     }
 
     /**
-     * Returns 'method' list
+     * Clears 'value' list
      *
-     * @return \Google\Protobuf\Compiler\MethodDescriptorProto[]
+     * @return null
      */
-    public function getMethod()
+    public function clearValue()
     {
-        return $this->get(self::METHOD);
+        return $this->clear(self::VALUE);
     }
 
     /**
-     * Returns 'method' iterator
+     * Returns 'value' list
+     *
+     * @return \Google\Protobuf\EnumValueDescriptorProto[]
+     */
+    public function getValue()
+    {
+        return $this->get(self::VALUE);
+    }
+
+    /**
+     * Returns 'value' iterator
      *
      * @return \ArrayIterator
      */
-    public function getMethodIterator()
+    public function getValueIterator()
     {
-        return new \ArrayIterator($this->get(self::METHOD));
+        return new \ArrayIterator($this->get(self::VALUE));
     }
 
     /**
-     * Returns element from 'method' list at given offset
+     * Returns element from 'value' list at given offset
      *
      * @param int $offset Position in list
      *
-     * @return \Google\Protobuf\Compiler\MethodDescriptorProto
+     * @return \Google\Protobuf\EnumValueDescriptorProto
      */
-    public function getMethodAt($offset)
+    public function getValueAt($offset)
     {
-        return $this->get(self::METHOD, $offset);
+        return $this->get(self::VALUE, $offset);
     }
 
     /**
-     * Returns count of 'method' list
+     * Returns count of 'value' list
      *
      * @return int
      */
-    public function getMethodCount()
+    public function getValueCount()
     {
-        return $this->count(self::METHOD);
+        return $this->count(self::VALUE);
     }
 
     /**
      * Sets value of 'options' property
      *
-     * @param \Google\Protobuf\Compiler\ServiceOptions $value Property value
+     * @param \Google\Protobuf\EnumOptions $value Property value
      *
      * @return null
      */
-    public function setOptions(\Google\Protobuf\Compiler\ServiceOptions $value)
+    public function setOptions(\Google\Protobuf\EnumOptions $value)
     {
         return $this->set(self::OPTIONS, $value);
     }
@@ -164,7 +166,7 @@ class ServiceDescriptorProto extends \ProtobufMessage
     /**
      * Returns value of 'options' property
      *
-     * @return \Google\Protobuf\Compiler\ServiceOptions
+     * @return \Google\Protobuf\EnumOptions
      */
     public function getOptions()
     {
