+++ src/Google/Protobuf/EnumOptions.php
@@ -1,30 +1,38 @@
 <?php
 /**
- * Auto generated from descriptor.proto at 2016-06-26 19:19:31
+ * Auto generated from descriptor.proto at 2016-07-08 14:22:37
+ *
+ * google.protobuf package
  */
 
-namespace Google\Protobuf\Compiler {
+namespace Google\Protobuf {
 /**
- * EnumValueOptions message
+ * EnumOptions message
  */
-class EnumValueOptions extends \ProtobufMessage
+class EnumOptions extends \ProtobufMessage
 {
     /* Field index constants */
-    const DEPRECATED = 1;
+    const ALLOW_ALIAS = 2;
+    const DEPRECATED = 3;
     const UNINTERPRETED_OPTION = 999;
 
     /* @var array Field descriptors */
     protected static $fields = array(
+        self::ALLOW_ALIAS => array(
+            'name' => 'allow_alias',
+            'required' => false,
+            'type' => \ProtobufMessage::PB_TYPE_BOOL,
+        ),
         self::DEPRECATED => array(
             'default' => false,
             'name' => 'deprecated',
             'required' => false,
-            'type' => 8,
+            'type' => \ProtobufMessage::PB_TYPE_BOOL,
         ),
         self::UNINTERPRETED_OPTION => array(
             'name' => 'uninterpreted_option',
             'repeated' => true,
-            'type' => '\Google\Protobuf\Compiler\UninterpretedOption'
+            'type' => '\Google\Protobuf\UninterpretedOption'
         ),
     );
 
@@ -43,6 +51,7 @@ class EnumValueOptions extends \ProtobufMessage
      */
     public function reset()
     {
+        $this->values[self::ALLOW_ALIAS] = null;
         $this->values[self::DEPRECATED] = self::$fields[self::DEPRECATED]['default'];
         $this->values[self::UNINTERPRETED_OPTION] = array();
     }
@@ -55,6 +64,28 @@ class EnumValueOptions extends \ProtobufMessage
     public function fields()
     {
         return self::$fields;
+    }
+
+    /**
+     * Sets value of 'allow_alias' property
+     *
+     * @param boolean $value Property value
+     *
+     * @return null
+     */
+    public function setAllowAlias($value)
+    {
+        return $this->set(self::ALLOW_ALIAS, $value);
+    }
+
+    /**
+     * Returns value of 'allow_alias' property
+     *
+     * @return boolean
+     */
+    public function getAllowAlias()
+    {
+        return $this->get(self::ALLOW_ALIAS);
     }
 
     /**
@@ -82,11 +113,11 @@ class EnumValueOptions extends \ProtobufMessage
     /**
      * Appends value to 'uninterpreted_option' list
      *
-     * @param \Google\Protobuf\Compiler\UninterpretedOption $value Value to append
+     * @param \Google\Protobuf\UninterpretedOption $value Value to append
      *
      * @return null
      */
-    public function appendUninterpretedOption(\Google\Protobuf\Compiler\UninterpretedOption $value)
+    public function appendUninterpretedOption(\Google\Protobuf\UninterpretedOption $value)
     {
         return $this->append(self::UNINTERPRETED_OPTION, $value);
     }
@@ -104,7 +135,7 @@ class EnumValueOptions extends \ProtobufMessage
     /**
      * Returns 'uninterpreted_option' list
      *
-     * @return \Google\Protobuf\Compiler\UninterpretedOption[]
+     * @return \Google\Protobuf\UninterpretedOption[]
      */
     public function getUninterpretedOption()
     {
@@ -126,7 +157,7 @@ class EnumValueOptions extends \ProtobufMessage
      *
      * @param int $offset Position in list
      *
-     * @return \Google\Protobuf\Compiler\UninterpretedOption
+     * @return \Google\Protobuf\UninterpretedOption
      */
     public function getUninterpretedOptionAt($offset)
     {
