+++ src/Google/Protobuf/FieldOptions.php
@@ -1,9 +1,11 @@
 <?php
 /**
- * Auto generated from descriptor.proto at 2016-06-26 19:19:31
+ * Auto generated from descriptor.proto at 2016-07-08 14:22:37
+ *
+ * google.protobuf package
  */
 
-namespace Google\Protobuf\Compiler {
+namespace Google\Protobuf {
 /**
  * FieldOptions message
  */
@@ -21,44 +23,44 @@ class FieldOptions extends \ProtobufMessage
     /* @var array Field descriptors */
     protected static $fields = array(
         self::CTYPE => array(
-            'default' => \Google\Protobuf\Compiler\FieldOptions_CType::STRING,
+            'default' => \Google\Protobuf\FieldOptions_CType::STRING,
             'name' => 'ctype',
             'required' => false,
-            'type' => 5,
+            'type' => \ProtobufMessage::PB_TYPE_INT,
         ),
         self::PACKED => array(
             'name' => 'packed',
             'required' => false,
-            'type' => 8,
+            'type' => \ProtobufMessage::PB_TYPE_BOOL,
         ),
         self::JSTYPE => array(
-            'default' => \Google\Protobuf\Compiler\FieldOptions_JSType::JS_NORMAL,
+            'default' => \Google\Protobuf\FieldOptions_JSType::JS_NORMAL,
             'name' => 'jstype',
             'required' => false,
-            'type' => 5,
+            'type' => \ProtobufMessage::PB_TYPE_INT,
         ),
         self::LAZY => array(
             'default' => false,
             'name' => 'lazy',
             'required' => false,
-            'type' => 8,
+            'type' => \ProtobufMessage::PB_TYPE_BOOL,
         ),
         self::DEPRECATED => array(
             'default' => false,
             'name' => 'deprecated',
             'required' => false,
-            'type' => 8,
+            'type' => \ProtobufMessage::PB_TYPE_BOOL,
         ),
         self::WEAK => array(
             'default' => false,
             'name' => 'weak',
             'required' => false,
-            'type' => 8,
+            'type' => \ProtobufMessage::PB_TYPE_BOOL,
         ),
         self::UNINTERPRETED_OPTION => array(
             'name' => 'uninterpreted_option',
             'repeated' => true,
-            'type' => '\Google\Protobuf\Compiler\UninterpretedOption'
+            'type' => '\Google\Protobuf\UninterpretedOption'
         ),
     );
 
@@ -231,11 +233,11 @@ class FieldOptions extends \ProtobufMessage
     /**
      * Appends value to 'uninterpreted_option' list
      *
-     * @param \Google\Protobuf\Compiler\UninterpretedOption $value Value to append
+     * @param \Google\Protobuf\UninterpretedOption $value Value to append
      *
      * @return null
      */
-    public function appendUninterpretedOption(\Google\Protobuf\Compiler\UninterpretedOption $value)
+    public function appendUninterpretedOption(\Google\Protobuf\UninterpretedOption $value)
     {
         return $this->append(self::UNINTERPRETED_OPTION, $value);
     }
@@ -253,7 +255,7 @@ class FieldOptions extends \ProtobufMessage
     /**
      * Returns 'uninterpreted_option' list
      *
-     * @return \Google\Protobuf\Compiler\UninterpretedOption[]
+     * @return \Google\Protobuf\UninterpretedOption[]
      */
     public function getUninterpretedOption()
     {
@@ -275,7 +277,7 @@ class FieldOptions extends \ProtobufMessage
      *
      * @param int $offset Position in list
      *
-     * @return \Google\Protobuf\Compiler\UninterpretedOption
+     * @return \Google\Protobuf\UninterpretedOption
      */
     public function getUninterpretedOptionAt($offset)
     {
