+++ src/Google/Protobuf/ServiceOptions.php
@@ -1,36 +1,32 @@
 <?php
 /**
- * Auto generated from descriptor.proto at 2016-06-26 19:19:31
+ * Auto generated from descriptor.proto at 2016-07-08 14:22:37
+ *
+ * google.protobuf package
  */
 
-namespace Google\Protobuf\Compiler {
+namespace Google\Protobuf {
 /**
- * EnumOptions message
+ * ServiceOptions message
  */
-class EnumOptions extends \ProtobufMessage
+class ServiceOptions extends \ProtobufMessage
 {
     /* Field index constants */
-    const ALLOW_ALIAS = 2;
-    const DEPRECATED = 3;
+    const DEPRECATED = 33;
     const UNINTERPRETED_OPTION = 999;
 
     /* @var array Field descriptors */
     protected static $fields = array(
-        self::ALLOW_ALIAS => array(
-            'name' => 'allow_alias',
-            'required' => false,
-            'type' => 8,
-        ),
         self::DEPRECATED => array(
             'default' => false,
             'name' => 'deprecated',
             'required' => false,
-            'type' => 8,
+            'type' => \ProtobufMessage::PB_TYPE_BOOL,
         ),
         self::UNINTERPRETED_OPTION => array(
             'name' => 'uninterpreted_option',
             'repeated' => true,
-            'type' => '\Google\Protobuf\Compiler\UninterpretedOption'
+            'type' => '\Google\Protobuf\UninterpretedOption'
         ),
     );
 
@@ -49,7 +45,6 @@ class EnumOptions extends \ProtobufMessage
      */
     public function reset()
     {
-        $this->values[self::ALLOW_ALIAS] = null;
         $this->values[self::DEPRECATED] = self::$fields[self::DEPRECATED]['default'];
         $this->values[self::UNINTERPRETED_OPTION] = array();
     }
@@ -62,28 +57,6 @@ class EnumOptions extends \ProtobufMessage
     public function fields()
     {
         return self::$fields;
-    }
-
-    /**
-     * Sets value of 'allow_alias' property
-     *
-     * @param boolean $value Property value
-     *
-     * @return null
-     */
-    public function setAllowAlias($value)
-    {
-        return $this->set(self::ALLOW_ALIAS, $value);
-    }
-
-    /**
-     * Returns value of 'allow_alias' property
-     *
-     * @return boolean
-     */
-    public function getAllowAlias()
-    {
-        return $this->get(self::ALLOW_ALIAS);
     }
 
     /**
@@ -111,11 +84,11 @@ class EnumOptions extends \ProtobufMessage
     /**
      * Appends value to 'uninterpreted_option' list
      *
-     * @param \Google\Protobuf\Compiler\UninterpretedOption $value Value to append
+     * @param \Google\Protobuf\UninterpretedOption $value Value to append
      *
      * @return null
      */
-    public function appendUninterpretedOption(\Google\Protobuf\Compiler\UninterpretedOption $value)
+    public function appendUninterpretedOption(\Google\Protobuf\UninterpretedOption $value)
     {
         return $this->append(self::UNINTERPRETED_OPTION, $value);
     }
@@ -133,7 +106,7 @@ class EnumOptions extends \ProtobufMessage
     /**
      * Returns 'uninterpreted_option' list
      *
-     * @return \Google\Protobuf\Compiler\UninterpretedOption[]
+     * @return \Google\Protobuf\UninterpretedOption[]
      */
     public function getUninterpretedOption()
     {
@@ -155,7 +128,7 @@ class EnumOptions extends \ProtobufMessage
      *
      * @param int $offset Position in list
      *
-     * @return \Google\Protobuf\Compiler\UninterpretedOption
+     * @return \Google\Protobuf\UninterpretedOption
      */
     public function getUninterpretedOptionAt($offset)
     {
